+++ cli/CHANGELOG.md
@@ -4,6 +4,18 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.36.5](https://github.com/wundergraph/cosmo/compare/wgc@0.36.4...wgc@0.36.5) (2024-01-09)
+
+**Note:** Version bump only for package wgc
+
+## [0.36.4](https://github.com/wundergraph/cosmo/compare/wgc@0.36.3...wgc@0.36.4) (2024-01-09)
+
+**Note:** Version bump only for package wgc
+
+## [0.36.3](https://github.com/wundergraph/cosmo/compare/wgc@0.36.2...wgc@0.36.3) (2024-01-08)
+
+**Note:** Version bump only for package wgc
+
 ## [0.36.2](https://github.com/wundergraph/cosmo/compare/wgc@0.36.1...wgc@0.36.2) (2024-01-06)
 
 **Note:** Version bump only for package wgc