+++ shared/CHANGELOG.md
@@ -4,6 +4,20 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.16.0](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.15.5...@wundergraph/cosmo-shared@0.16.0) (2024-01-09)
+
+### Features
+
+* add support of interface objects ([#407](https://github.com/wundergraph/cosmo/issues/407)) ([3d7b0e1](https://github.com/wundergraph/cosmo/commit/3d7b0e1f55fd8087945923a8e4f5e7d66f6b559a)) (@Aenimus)
+
+## [0.15.5](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.15.4...@wundergraph/cosmo-shared@0.15.5) (2024-01-09)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
+## [0.15.4](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.15.3...@wundergraph/cosmo-shared@0.15.4) (2024-01-08)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
 ## [0.15.3](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.15.2...@wundergraph/cosmo-shared@0.15.3) (2024-01-06)
 
 **Note:** Version bump only for package @wundergraph/cosmo-shared